% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_columns.R
\name{drop_columns}
\alias{drop_columns}
\title{drop_columns}
\usage{
drop_columns(x, data)
}
\arguments{
\item{x}{The percentage (as a decimal) of missing values allowed in a column}

\item{data}{The dataset on which to remove columns from}
}
\value{
Returns the dataset with columns having \code{x}\% or more missing values removed of
}
\description{
\code{drop_columns} gets rid of columns from a dataset that have large amounts of missing values.
}
\examples{

new_data <- drop_columns(0.9, ratings_wide)
# returns our movie ratings dataset with columns having over 90\% missing values removed

}
\references{
https://stackoverflow.com/questions/31848156/delete-columns-rows-with-more-than-x-missing
}
