% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_rows.R
\name{drop_rows}
\alias{drop_rows}
\title{drop_rows}
\usage{
drop_rows(y, data)
}
\arguments{
\item{y}{The percentage (as a decimal) of missing values allowed in a row}

\item{data}{The dataset on which to remove rows from}
}
\value{
Returns the dataset with rows having \code{y}\% or more missing values removed of
}
\description{
\code{drop_rows} gets rid of rows from a dataset that have large amounts of missing values.
}
\examples{

new_data <- drop_rows(0.75, new_data) # returns our movie ratings dataset with rows having over 75\% missing values removed

}
