% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knn_clustered_data.R
\name{get_knn_clustered_data}
\alias{get_knn_clustered_data}
\title{get_knn_clustered_data}
\usage{
get_knn_clustered_data(original, data)
}
\arguments{
\item{original}{an original dataset to get row ids from}

\item{data}{a dataset of knn imputed values}
}
\value{
Returns the clustered dataset of kmean clusters using knn
}
\description{
\code{get_knn_clustered_data} returns the clustered dataset of kmeans clusters for knn
}
\examples{

knn_clustered_data <- get_knn_clustered_data(new_data, knn_imputed) # returns kmeans 2 cluster plot for knn on our movie ratings data

}
