% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knn_kmeans_elbow_plot.R
\name{get_knn_kmeans_elbow_plot}
\alias{get_knn_kmeans_elbow_plot}
\title{get_knn_kmeans_elbow_plot}
\usage{
get_knn_kmeans_elbow_plot(data)
}
\arguments{
\item{data}{A dataset of knn imputed values to find the optimal number of kmeans clusters on}
}
\value{
Returns the elbow plot of optimal kmeans clusters
}
\description{
\code{get_knn_kmeans_elbow_plot} returns the elbow plot of kmeans clusters for knn
}
\examples{

get_knn_kmeans_elbow_plot(knn_imputed) # returns elbow plot for knn on our movie ratings data

}
