% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nmf_clustered_data.R
\name{get_nmf_clustered_data}
\alias{get_nmf_clustered_data}
\title{get_nmf_clustered_data}
\usage{
get_nmf_clustered_data(nmf, kmean)
}
\arguments{
\item{nmf}{NMF}

\item{kmean}{kmean}
}
\value{
Returns the clustered dataset of kmean clusters using NMF
}
\description{
\code{get_nmf_clustered_data} returns the clustered dataset of kmeans clusters for NMF
}
\examples{

nmf_clustered_data <- get_nmf_clustered_data(nmf, kmean)
# returns kmeans clustered dataset of movies using NMF on our movie ratings data

}
