% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nmf_kmeans_elbow_plot.R
\name{get_nmf_kmeans_elbow_plot}
\alias{get_nmf_kmeans_elbow_plot}
\title{get_nmf_kmeans_elbow_plot}
\usage{
get_nmf_kmeans_elbow_plot(nmf)
}
\arguments{
\item{nmf}{NMF}
}
\value{
Returns the elbow plot of optimal kmeans clusters
}
\description{
\code{get_nmf_kmeans_elbow_plot} returns the elbow plot of kmeans clusters for NMF
}
\examples{

get_nmf_kmeans_elbow_plot(nmf) # returns elbow plot for NMF on our movie ratings data

}
