% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knn_imputed}
\alias{knn_imputed}
\title{knn_imputed}
\format{
\subsection{\code{knn_imputed}}{

Matrix 138x317 of knn imputed movie ratings:
}
}
\usage{
knn_imputed
}
\description{
A matrix of knn imputed imputed values
}
\keyword{datasets}
