% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movie_titles}
\alias{movie_titles}
\title{movie_titles}
\format{
\subsection{\code{movie_titles}}{

A dataframe with 9,742 rows and 3 columns:
\describe{
\item{movieId}{Unique movie Id}
\item{title}{Movie Titke}
\item{genres}{Movie genres}
...
}
}
}
\source{
\url{https://grouplens.org/datasets/movielens/}
}
\usage{
movie_titles
}
\description{
A dataset of movie titles from MovieLens
}
\keyword{datasets}
