% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovieDataCleaner-package.R
\docType{package}
\name{MovieDataCleaner-package}
\alias{MovieDataCleaner}
\alias{MovieDataCleaner-package}
\title{MovieDataCleaner: Cleaning Datasets with KNN and NMF Imputation}
\description{
This package is able to impute missing values into datasets with large amounts of missing values using both knn imputation and NMF imputation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BruceBuurstra/MovieDataCleaner}
  \item Report bugs at \url{https://github.com/BruceBuurstra/MovieDataCleaner/issues}
}

}
\author{
\strong{Maintainer}: Bruce Buurstra \email{buursbru@mail.gvsu.edu}

Authors:
\itemize{
  \item Madan K C
  \item Rashmita Vagu
  \item Durga saranya tummalapalli
}

}
\keyword{internal}
