% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_data}
\alias{new_data}
\title{new_data}
\format{
\subsection{\code{new_data}}{

A dataframe with 138 rows (users) and 317 columns (movies):
\describe{
\item{rows}{userId}
\item{columns}{movieId}
\item{data}{movie ratings}
...
}
}
}
\usage{
new_data
}
\description{
A subset of ratings_wide data
}
\keyword{datasets}
