% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_knn_kmeans_clusters.R
\name{optimal_knn_kmeans_clusters}
\alias{optimal_knn_kmeans_clusters}
\title{optimal_knn_kmeans_clusters}
\usage{
optimal_knn_kmeans_clusters(data)
}
\arguments{
\item{data}{A dataset of knn imputed values to find the optimal number of kmeans clusters on}
}
\value{
Returns the plot of optimal kmeans clusters using average silhouette width
}
\description{
\code{optimal_knn_kmeans_clusters} finds the optimal number of kmeans clusters for knn
}
\examples{

optimal_knn_kmeans_clusters(knn_imputed) # returns plot for knn on our movie ratings data

}
