% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_nmf_kmeans_clusters.R
\name{optimal_nmf_kmeans_clusters}
\alias{optimal_nmf_kmeans_clusters}
\title{optimal_nmf_kmeans_clusters}
\usage{
optimal_nmf_kmeans_clusters(nmf)
}
\arguments{
\item{nmf}{NMF}
}
\value{
Returns the plot of optimal kmeans clusters using average silhouette width
}
\description{
\code{optimal_nmf_kmeans_clusters} finds the optimal number of kmeans clusters for NMF
}
\examples{

optimal_nmf_kmeans_clusters(nmf) # returns plot for NMF on our movie ratings data

}
