% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratings_wide}
\alias{ratings_wide}
\title{ratings_wide}
\format{
\subsection{\code{ratings_wide}}{

A dataframe with 610 rows (users) and 9725 columns (movies):
\describe{
\item{rows}{userId}
\item{columns}{movieId}
\item{data}{movie ratings}
...
}
}
}
\source{
\url{https://grouplens.org/datasets/movielens/}
}
\usage{
ratings_wide
}
\description{
A  dataset of movie ratings from MovieLens
}
\keyword{datasets}
