% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to_movie_rating.R
\name{round_to_movie_rating}
\alias{round_to_movie_rating}
\title{round_to_movie_rating}
\usage{
round_to_movie_rating(x)
}
\arguments{
\item{x}{A value}
}
\value{
Returns the value of \code{x} rounded to a movie rating
}
\description{
\code{round_to_movie_rating} gets rid of rows from a dataset that have large amounts of missing values.
}
\examples{

round_to_movie_rating(4.41) # returns 4.5

}
