% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_knn_imputation.R
\name{run_knn_imputation}
\alias{run_knn_imputation}
\title{run_knn_imputation}
\usage{
run_knn_imputation(n, data)
}
\arguments{
\item{n}{the k amount of nearest neighbors to use}

\item{data}{A dataset to run knn on}
}
\value{
Returns the dataset using knn to impute missing values
}
\description{
\code{run_knn_imputation} imputes the missing values into a dataset using knn
}
\examples{

run_knn_imputation(5, new_data)
# returns our movie ratings dataset with missing values imputed from using knn of 5

}
