% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_knn_movies_clustering.R
\name{run_knn_movies_clustering}
\alias{run_knn_movies_clustering}
\title{run_knn_movies_clustering}
\usage{
run_knn_movies_clustering(v, data)
}
\arguments{
\item{v}{number of clusters}

\item{data}{A dataset of knn imputed values}
}
\value{
Returns the clustered plot of \code{v} kmeans clusters
}
\description{
\code{run_knn_movies_clustering} returns the clustered plot of kmeans clusters for knn
}
\examples{

run_knn_movies_clustering(2, knn_imputed)
# returns kmeans 2 cluster plot for knn on our movie ratings data

}
