% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nmf_imputation.R
\name{run_nmf_imputation}
\alias{run_nmf_imputation}
\title{run_nmf_imputation}
\usage{
run_nmf_imputation(z, data)
}
\arguments{
\item{z}{The rank of the NMF}

\item{data}{A dataset to run NMF on}
}
\value{
Returns the dataset using NMF of rank \code{z} to impute missing values
}
\description{
\code{run_nmf_imputation} reduces the dimension of dataset using nmf
}
\examples{

 # returns our movie ratings dataset with missing values imputed from an NMF of rank 3

}
