% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nmf_movies_clustering.R
\name{run_nmf_movies_clustering}
\alias{run_nmf_movies_clustering}
\title{run_nmf_movies_clustering}
\usage{
run_nmf_movies_clustering(w, nmf)
}
\arguments{
\item{w}{number of clusters}

\item{nmf}{NMF}
}
\value{
Returns the clustered plot of \code{w} kmeans clusters
}
\description{
\code{run_nmf_movies_clustering} returns the clustered plot of kmeans clusters for NMF
}
\examples{

run_nmf_movies_clustering(2, nmf)
# returns kmeans 2 cluster plot for NMF on our movie ratings data

}
